+++ CHANGELOG.rst
@@ -7,12 +7,30 @@ smaht-portal
 Change Log
 ----------
 
-0.37.0
+0.37.2
 ======
 
 `PR 131: feat: Benchmarking Page navigation toggle functionality <https://github.com/smaht-dac/smaht-portal/pull/131>`_
 
 * Enable toggle for benchmarking page navigation
+
+
+0.37.1
+======
+
+`PR 132: Fix file embeds <https://github.com/smaht-dac/smaht-portal/pull/132>`_
+
+* Fix file embeds by removing sub-type LinkTos from Samples to SampleSources and deletion of unused embed
+
+
+0.37.0
+======
+
+`PR 127: File Overview Calcprops + Item Utils <https://github.com/smaht-dac/smaht-portal/pull/127>`_
+
+* Add calcprops for file overview page
+* Add calcprops for associated items on files
+* Add item_utils module for common item functions
 
 
 0.36.0